Save netcdf file (#113)
* added save netcdf file modal; still need to add backend functionality

* added backend functionality

* fixed save button label, close button functionality, and state resetting

* added props for deflate and shuffle to codeInjector

* changed deflateValue type to number

* added alerts when saving a netCDF file

* code cleanup

* fixed bug that was introduced during rebasing on master branchthat removed selectedVariables from state

* fixed export alert when saving Variable

* fixed errors caused by merging

* changed click method in tests

* added a slight delay right after clicking on 'Update' button on the 'Edit Axis' pop up

* renamed button from "save variable" to "save" to use less space

* fixed bug caused by merge

* Added rename functionality to edit modal of variable.

* Fixed bugs related to variable rename feature. Updated delete button to bring up pop-up warning. Changed the varLoader variable rename feature so name is updated as user types. Updated save dialog so that selection/deselection is off when the dialog is open. Fixed minor bug where deleting a variable would not deselect it.

* Updated validation logic in varLoader and varMini so that renaming a variable is faster. The varMini now includes a warning if user is attempting to rename variable to an existing variable.

* Updated a selenium test button locator.

* Fixed some bugs with the save variable feature. Cleaned up commented out code.

* Updated edit_axis locator

* Another fix

* Modified test locator.

* Updated edit variable form to make sure variable is not deselected after it is edited.

+++ src/components/VarMini.tsx
@@ -1,10 +1,18 @@
 // Dependencies
 import * as React from "react";
+import { NotebookPanel } from "@jupyterlab/notebook";
 import {
+  Alert,
   Badge,
   Button,
   Card,
   CardBody,
+  Collapse,
+  CustomInput,
+  Input,
+  InputGroup,
+  InputGroupAddon,
+  Label,
   Modal,
   ModalBody,
   ModalFooter,
@@ -12,10 +20,13 @@ import {
 } from "reactstrap";
 
 // Project Components
+import { CodeInjector } from "../CodeInjector";
 import { Utilities } from "../Utilities";
 import { AxisInfo } from "./AxisInfo";
 import { DimensionSlider } from "./DimensionSlider";
 import { Variable } from "./Variable";
+import { NotebookUtilities } from "../NotebookUtilities";
+import { checkForExportedFileCommand } from "../PythonCommands";
 import { ISignal } from "@phosphor/signaling";
 import { VariableTracker } from "../VariableTracker";
 
@@ -27,32 +38,73 @@ const badgeStyle: React.CSSProperties = {
   marginLeft: "0.5em"
 };
 
+const modalOverflow: React.CSSProperties = {
+  maxHeight: "70vh",
+  overflow: "auto"
+};
+
+// Regex filter for unallowed variable names
+const forbidden: RegExp = /^[^a-z_]|[^a-z0-9_]+/i;
+
 interface IVarMiniProps {
   buttonColor: string; // The hex value for the color
+  codeInjector: CodeInjector;
   variable: Variable; // the variable this component will show
   varSelectionChanged: ISignal<VariableTracker, string[]>;
-  updateDimInfo: (newInfo: any, varID: string) => void; // method passed by the parent to update their copy of the variables dimension info
+  varAliasExists: (varAlias: string) => boolean; // Method that returns true if specified variable name is already taken
   isSelected: (alias: string) => boolean; // method to check if this variable is selected in parent
   selected: boolean; // should the axis be hidden by default
-  copyVariable: (variable: Variable, newName: string) => Promise<void>;
+  copyVariable: (
+    variable: Variable,
+    newName: string,
+    addVariable: boolean
+  ) => Variable;
   deleteVariable: (variable: Variable) => Promise<void>;
+  selectVariable: (varID: string) => Promise<void>;
   modalOpen: (isOpen: boolean) => void;
   selectOrder: number;
   allowReload: boolean; // is this variable allowed to be reloaded
-  reload: () => void; // a function to reload the variable
+  reload: (item: Variable, newAlias?: string) => void; // a function to reload the variable
+  setPlotInfo: (plotName: string, plotFormat: string) => void;
+  exportAlerts: () => void;
+  dismissSavePlotSpinnerAlert: () => void;
+  showExportSuccessAlert: () => void;
+  notebookPanel: NotebookPanel;
 }
 interface IVarMiniState {
-  showAxis: boolean; // should the edit axis modal be shown
+  activateAppend: boolean;
+  activateShuffle: boolean;
+  activateDeflate: boolean;
+  deflateValue: number;
+  filename: string;
+  nameState: NAME_STATUS;
+  nameValue: string;
+  newVariableSaveName: string;
   selected: boolean;
+  showAxis: boolean; // should the edit axis modal be shown
+  showSaveModal: boolean;
+  validateFileName: boolean;
   variable: Variable;
 }
+
+export type NAME_STATUS = "Invalid!" | "Variable Already Exists!" | "Valid";
 
 export class VarMini extends React.Component<IVarMiniProps, IVarMiniState> {
   constructor(props: IVarMiniProps) {
     super(props);
     this.state = {
+      activateAppend: false,
+      activateDeflate: false,
+      activateShuffle: false,
+      deflateValue: 0,
+      filename: "",
+      nameState: "Valid",
+      nameValue: this.props.variable.alias,
+      newVariableSaveName: "",
       selected: props.selected,
       showAxis: false,
+      showSaveModal: false,
+      validateFileName: false,
       variable: this.props.variable
     };
     this.openMenu = this.openMenu.bind(this);
@@ -61,7 +113,20 @@ export class VarMini extends React.Component<IVarMiniProps, IVarMiniState> {
     this.handleDeleteClick = this.handleDeleteClick.bind(this);
     this.handleUpdateClick = this.handleUpdateClick.bind(this);
     this.handleCopyClick = this.handleCopyClick.bind(this);
+    this.handleSaveClick = this.handleSaveClick.bind(this);
+    this.toggleSaveModal = this.toggleSaveModal.bind(this);
+    this.toggleShuffle = this.toggleShuffle.bind(this);
+    this.toggleDeflate = this.toggleDeflate.bind(this);
+    this.toggleAppend = this.toggleAppend.bind(this);
+    this.onFilenameChange = this.onFilenameChange.bind(this);
+    this.updateDeflateValue = this.updateDeflateValue.bind(this);
+    this.updateNewVariableName = this.updateNewVariableName.bind(this);
+    this.save = this.save.bind(this);
+    this.dismissFilenameValidation = this.dismissFilenameValidation.bind(this);
     this.updateSelections = this.updateSelections.bind(this);
+    this.handleNameInput = this.handleNameInput.bind(this);
+    this.validateNameInput = this.validateNameInput.bind(this);
+    this.updateDimensionInfo = this.updateDimensionInfo.bind(this);
   }
 
   public componentDidMount(): void {
@@ -72,7 +137,25 @@ export class VarMini extends React.Component<IVarMiniProps, IVarMiniState> {
     this.props.varSelectionChanged.disconnect(this.updateSelections);
   }
 
-  public updateSelections() {
+  public reset(): void {
+    this.setState({
+      activateAppend: false,
+      activateDeflate: false,
+      activateShuffle: false,
+      deflateValue: 0,
+      filename: "",
+      nameState: "Valid",
+      nameValue: this.props.variable.alias,
+      newVariableSaveName: "",
+      selected: this.props.selected,
+      showAxis: false,
+      showSaveModal: false,
+      validateFileName: false,
+      variable: this.props.variable
+    });
+  }
+
+  public updateSelections(): void {
     this.setState({
       selected: this.props.isSelected(this.state.variable.varID)
     });
@@ -99,7 +182,132 @@ export class VarMini extends React.Component<IVarMiniProps, IVarMiniState> {
     });
   }
 
+  /**
+   * @description Toggles the save variable modal
+   */
+  public toggleSaveModal(): void {
+    if (this.state.showSaveModal) {
+      this.props.modalOpen(false);
+    } else {
+      this.props.modalOpen(true);
+    }
+    this.setState({
+      activateDeflate: false,
+      activateShuffle: false,
+      deflateValue: 0,
+      filename: "",
+      showSaveModal: !this.state.showSaveModal
+    });
+  }
+
+  /**
+   * @description Toggles the shuffle switch
+   */
+  public toggleShuffle(): void {
+    this.setState({
+      activateShuffle: !this.state.activateShuffle
+    });
+  }
+
+  /**
+   * @description Toggles the deflate switch
+   */
+  public toggleDeflate(): void {
+    this.setState(
+      {
+        activateDeflate: !this.state.activateDeflate
+      },
+      () => {
+        if (!this.state.activateDeflate) {
+          this.setState({
+            deflateValue: 0
+          });
+        }
+      }
+    );
+  }
+
+  /**
+   * @description Toggles the deflate switch
+   */
+  public toggleAppend(): void {
+    this.setState({
+      activateAppend: !this.state.activateAppend
+    });
+  }
+
+  public onFilenameChange(event: React.ChangeEvent<HTMLInputElement>) {
+    this.setState({ filename: event.target.value });
+  }
+
+  public updateDeflateValue(event: React.ChangeEvent<HTMLInputElement>) {
+    this.setState({ deflateValue: parseInt(event.target.value, 10) });
+  }
+
+  public updateNewVariableName(event: React.ChangeEvent<HTMLInputElement>) {
+    this.setState({ newVariableSaveName: event.target.value });
+  }
+
+  public dismissFilenameValidation() {
+    this.setState({ validateFileName: false });
+  }
+
+  public async save() {
+    const splitFileName = this.state.filename.split(".");
+
+    if (!this.state.filename) {
+      this.setState({ validateFileName: true });
+      return;
+    }
+    this.setState({ validateFileName: false });
+    await this.props.codeInjector.saveNetCDFFile(
+      this.state.filename,
+      this.state.variable.alias,
+      this.state.newVariableSaveName,
+      this.state.activateAppend,
+      this.state.activateShuffle,
+      this.state.activateDeflate,
+      this.state.deflateValue
+    );
+    this.toggleSaveModal();
+    this.props.setPlotInfo(splitFileName[0], splitFileName[1]);
+    this.props.exportAlerts();
+
+    try {
+      const result: string = await NotebookUtilities.sendSimpleKernelRequest(
+        this.props.notebookPanel,
+        checkForExportedFileCommand(this.state.filename)
+      );
+      if (result === "True") {
+        this.props.dismissSavePlotSpinnerAlert();
+        this.props.showExportSuccessAlert();
+      }
+    } catch (error) {
+      console.error("error with checking file:", error);
+    }
+  }
+
+  public updateDimensionInfo(newInfo: any): void {
+    const updatedVar: Variable = this.state.variable;
+    updatedVar.axisInfo.forEach((axis: AxisInfo, axisIndex: number) => {
+      if (axis.name !== newInfo.name) {
+        return;
+      }
+      updatedVar.axisInfo[axisIndex].min = newInfo.min;
+      updatedVar.axisInfo[axisIndex].max = newInfo.max;
+    });
+    this.setState({ variable: updatedVar });
+  }
+
   public render(): JSX.Element {
+    // Set the input color
+    let nameStateColor = "success";
+    if (this.state.nameState === "Invalid!") {
+      nameStateColor = "danger";
+    } else if (this.state.nameState === "Variable Already Exists!") {
+      nameStateColor = "warning";
+    }
+
     return (
       <div>
         <div
@@ -111,7 +319,6 @@ export class VarMini extends React.Component<IVarMiniProps, IVarMiniState> {
             className={/*@tag<varmini-name-btn>*/ "varmini-name-btn-vcdat"}
             outline={true}
             color={"success"}
-            style={{ backgroundColor: this.props.buttonColor }}
             active={this.props.isSelected(this.state.variable.varID)}
           >
             {this.state.variable.alias}
@@ -122,13 +329,29 @@ export class VarMini extends React.Component<IVarMiniProps, IVarMiniState> {
             style={axisStyle}
             title={
               this.state.variable.sourceName === ""
-                ? "Editing of modified variables disabled for now."
+                ? "Note: This variable was not loaded from a file."
                 : ""
             }
-            color={"danger"}
+            color={"info"}
             onClick={this.handleEditClick}
           >
             edit
+          </Button>
+          <Button
+            outline={true}
+            color={"secondary"}
+            style={axisStyle}
+            onClick={this.handleSaveClick}
+          >
+            save
+          </Button>
+          <Button
+            outline={true}
+            color={"danger"}
+            style={axisStyle}
+            onClick={this.handleDeleteClick}
+          >
+            delete
           </Button>
           {this.props.isSelected(this.state.variable.varID) && (
             <Badge
@@ -148,17 +371,38 @@ export class VarMini extends React.Component<IVarMiniProps, IVarMiniState> {
           toggle={this.toggleModal}
           size="lg"
         >
-          <ModalHeader toggle={this.toggleModal}>Edit Axis</ModalHeader>
+          <ModalHeader toggle={this.toggleModal}>
+            Edit Variable: {this.state.variable.alias}
+          </ModalHeader>
           <ModalBody
             className={/*@tag<varmini-edit-modal>*/ "varmini-edit-modal-vcdat"}
           >
+            <Card>
+              <CardBody>
+                <InputGroup style={{ marginTop: "5px" }}>
+                  <InputGroupAddon addonType="prepend">
+                    Rename Variable:
+                  </InputGroupAddon>
+                  <Input
+                    onChange={this.handleNameInput}
+                    value={this.state.nameValue}
+                    placeholder="Enter new name here."
+                  />
+                  <InputGroupAddon addonType="append">
+                    <Button color={nameStateColor} disabled={true}>
+                      {this.state.nameState}
+                    </Button>
+                  </InputGroupAddon>
+                </InputGroup>
+              </CardBody>
+            </Card>
             {this.state.showAxis &&
               this.state.variable.axisInfo.length > 0 &&
               this.state.variable.axisInfo.map((item: AxisInfo) => {
                 if (!item.data || item.data.length <= 1) {
                   return;
                 }
-                item.updateDimInfo = this.props.updateDimInfo;
+                item.updateDimInfo = this.updateDimensionInfo;
                 return (
                   <div key={item.name} style={axisStyle}>
                     <Card>
@@ -185,6 +429,96 @@ export class VarMini extends React.Component<IVarMiniProps, IVarMiniState> {
             </Button>
           </ModalFooter>
         </Modal>
+        <Modal
+          id="var-save-modal"
+          isOpen={this.state.showSaveModal}
+          toggle={this.toggleSaveModal}
+          size="lg"
+        >
+          <ModalHeader toggle={this.toggleSaveModal}>Save Variable</ModalHeader>
+          <ModalBody style={modalOverflow}>
+            <Label>Filename:</Label>
+            <Input
+              type="text"
+              name="text"
+              placeholder="Name.nc"
+              value={this.state.filename}
+              onChange={this.onFilenameChange}
+            />
+            <Alert
+              color="danger"
+              isOpen={this.state.validateFileName}
+              toggle={this.dismissFilenameValidation}
+            >
+              The file name can not be blank
+            </Alert>
+            {/*<CustomInput
+              type="switch"
+              id="appendSwitch"
+              name="appendSwitch"
+              label="Append to existing file"
+              checked={this.state.activateAppend}
+              onChange={this.toggleAppend}
+            />
+            <br />*/}
+            <Label>Variable Name in File:</Label>
+            <Input
+              type="text"
+              name="text"
+              placeholder={this.state.variable.alias}
+              value={this.state.newVariableSaveName}
+              onChange={this.updateNewVariableName}
+            />
+            <br />
+            <CustomInput
+              type="switch"
+              id="shuffleSwitch"
+              name="shuffleSwitch"
+              label="Shuffle"
+              checked={this.state.activateShuffle}
+              onChange={this.toggleShuffle}
+            />
+            <br />
+            <CustomInput
+              type="switch"
+              id="deflateSwitch"
+              name="deflateSwitch"
+              label="Deflate"
+              checked={this.state.activateDeflate}
+              onChange={this.toggleDeflate}
+            />
+            <br />
+            <Collapse isOpen={this.state.activateDeflate}>
+              <Label>Deflate Level:</Label>
+              <Input
+                type="select"
+                name="select"
+                id="deflateSelect"
+                value={this.state.deflateValue}
+                onChange={this.updateDeflateValue}
+              >
+                <option>0</option>
+                <option>1</option>
+                <option>2</option>
+                <option>3</option>
+                <option>4</option>
+                <option>5</option>
+                <option>6</option>
+                <option>7</option>
+                <option>8</option>
+                <option>9</option>
+              </Input>
+            </Collapse>
+          </ModalBody>
+          <ModalFooter>
+            <Button color="primary" onClick={this.save}>
+              Save
+            </Button>{" "}
+            <Button color="secondary" onClick={this.toggleSaveModal}>
+              Cancel
+            </Button>
+          </ModalFooter>
+        </Modal>
       </div>
     );
   }
@@ -193,19 +527,32 @@ export class VarMini extends React.Component<IVarMiniProps, IVarMiniState> {
     clickEvent: React.MouseEvent<HTMLButtonElement>
   ): Promise<void> {
     clickEvent.stopPropagation();
-    await this.props.copyVariable(this.state.variable, "Test");
+    const copy: Variable = this.props.copyVariable(
+      this.state.variable,
+      this.state.nameValue,
+      true
+    );
+    await this.props.codeInjector.loadVariable(copy);
   }
 
   private async handleDeleteClick(
     clickEvent: React.MouseEvent<HTMLButtonElement>
   ): Promise<void> {
     clickEvent.stopPropagation();
-    await this.props.deleteVariable(this.state.variable);
+    const result: boolean = await NotebookUtilities.showYesNoDialog(
+      "Warning!",
+      "Are you sure you want to delete this variable?",
+      "OK",
+      "Cancel"
+    );
+    if (result) {
+      await this.props.deleteVariable(this.state.variable);
+    }
   }
 
-  private handleEditClick(
+  private async handleEditClick(
     clickEvent: React.MouseEvent<HTMLButtonElement>
-  ): void {
+  ): Promise<void> {
     clickEvent.stopPropagation();
     if (this.state.variable.axisInfo.length > 0) {
       this.setState({ showAxis: true });
@@ -216,8 +563,68 @@ export class VarMini extends React.Component<IVarMiniProps, IVarMiniState> {
     }
   }
 
-  private handleUpdateClick(): void {
+  private handleSaveClick(
+    clickEvent: React.MouseEvent<HTMLButtonElement>
+  ): void {
+    this.toggleSaveModal();
+    clickEvent.stopPropagation();
+  }
+
+  private validateNameInput(nameEntry: string): void {
+    if (nameEntry === this.state.variable.alias) {
+      this.setState({ nameState: "Valid" });
+      return;
+    }
+
+    const invalid: boolean = forbidden.test(nameEntry);
+    let state: NAME_STATUS = "Valid";
+    if (nameEntry === "" || invalid) {
+      state = "Invalid!";
+    } else if (this.props.varAliasExists(nameEntry)) {
+      state = "Variable Already Exists!";
+    }
+
+    this.setState({ nameState: state });
+  }
+
+  private handleNameInput(event: React.ChangeEvent<HTMLInputElement>): void {
+    const nameEntry: string = event.target.value;
+    this.setState({ nameValue: nameEntry });
+    this.validateNameInput(nameEntry);
+  }
+
+  private async handleUpdateClick(): Promise<void> {
+    if (this.state.nameState === "Variable Already Exists!") {
+      const proceed: boolean = await NotebookUtilities.showYesNoDialog(
+        "Warning!",
+        "Another variable already has this name. In order to rename this variable the other will be replaced. Continue?",
+        "Yes, replace other variable.",
+        "Cancel"
+      );
+      if (!proceed) {
+        return;
+      }
+    }
+
     this.toggleModal();
-    this.props.reload();
+    // Load new variable
+    await this.props.reload(this.state.variable, this.state.nameValue);
+
+    // Delete old variable if rename was done
+    if (this.state.variable.alias !== this.state.nameValue) {
+      // Create a copy with the new name
+      const copy: Variable = this.props.copyVariable(
+        this.state.variable,
+        this.state.nameValue,
+        true
+      );
+
+      // Delete variable of the old name
+      await this.props.deleteVariable(this.state.variable);
+      await this.setState({ variable: copy });
+    }
+    // Select the new variable (if previous one deleted)
+    this.props.selectVariable(this.state.variable.varID);
+    this.reset();
   }
 }
